Revert "1) Replaced startfield with ints."
This caused stars to slide unnaturally.

This reverts commit 488c5c9ac00e677a02f651dd4786aa4dc485cd85.

+++ Source/Engine/starfield.cpp
@@ -30,8 +30,8 @@ Starfield::Starfield( int num ) {
 	for( i = 0; i < num; i++ ) {
 		int c;
 
-		stars[i].x = (rand() % (int)(1.3 * Video::GetWidth()));
-		stars[i].y = (rand() % (int)(1.4 * Video::GetHeight()));
+		stars[i].x = (float)(rand() % (int)(1.3 * Video::GetWidth()));
+		stars[i].y = (float)(rand() % (int)(1.4 * Video::GetHeight()));
 		c = rand() % 225; // generate greys between 0 and 225
 		stars[i].clr = static_cast<float>( c / 256. );
 	}
@@ -61,21 +61,21 @@ void Starfield::Update( void ) {
 	int i;
 	double dx, dy;
 	Camera *camera = Camera::Instance();
-	short int w, h;
+	float w, h;
 
 	camera->GetDelta( &dx, &dy );
 	
-	w = 1.3 * Video::GetWidth();
-	h = 1.4 * Video::GetHeight();
+	w = static_cast<float>(1.3 * Video::GetWidth());
+	h = static_cast<float>(1.4 * Video::GetHeight());
 	
 	for( i = 0; i < num; i++ ) {
-		stars[i].x -= dx * stars[i].clr;
-		stars[i].y -= dy * stars[i].clr;
+		stars[i].x -= (float)dx * stars[i].clr;
+		stars[i].y -= (float)dy * stars[i].clr;
 
 		// handle wrapping the stars around if they go offscreen top/left
-		while( stars[i].x < 0 )
+		while( stars[i].x < 0.0f )
 			stars[i].x += w;
-		while( stars[i].y < 0 )
+		while( stars[i].y < 0.0f )
 			stars[i].y += h;
 
 		// handle wrapping the stars around if they go offscreen bottom/right
